[CHORE] 공유프로세스 알림 시간 Text Binding 방식 변경 (#415)
