+++ app/src/main/java/org/sopt/havit/ui/share/ContentsSummeryFragment.kt
@@ -75,26 +75,9 @@ class ContentsSummeryFragment : Fragment() {
     private fun getNotificationTime() {
         // 알림을 설정 했다면 tv_set_alarm 값을 알림설정한 시간으로 변경
         val setTime = MySharedPreference.getNotificationTime(requireContext())
-        binding.alarm = if (setTime.isNotEmpty()) setDateFormat(setTime) else "알림 설정"
+        binding.alarm = setTime.ifEmpty { getString(R.string.set_alarm_space) }
     }
 
-    private fun setDateFormat(originTime: String): String {
-        Log.d("originTime", originTime) // 2022.01.25 00:04:54
-
-        // 날짜 (2022.01.25)
-        val date =
-            "${originTime[2]}${originTime[3]}.${originTime[5]}${originTime[6]}.${originTime[8]}${originTime[9]}"
-        // 시 (오후 11시 :: 12시간제 적용)
-        val hour = "${originTime[11]}${originTime[12]}".toInt()
-        val newHour = when (hour) {
-            in 0..12 -> " 오전 ${hour}시 "
-            else -> " 오후 ${hour - 12}시 "
-        }
-        // 분 (3분 :: 자릿수 재졍렬을 위한 이중 형변환 사용)
-        val min = "${originTime[14]}${originTime[15]}".toInt().toString() + "분"
-
-        return "$date$newHour$min 알림 예정"
-    }
 
     private fun setContents() {
         val url = getUrl()
