+++ app/src/main/java/org/sopt/havit/util/BindingAdapter.kt
@@ -1,6 +1,7 @@
 package org.sopt.havit.util
 
 import android.graphics.drawable.Drawable
+import android.util.Log
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
@@ -120,12 +121,36 @@ object BindingAdapter {
 
     @BindingAdapter("descriptionImg")
     @JvmStatic
-    fun setDescriptionImg(imageView: ImageView,rate: Int){
+    fun setDescriptionImg(imageView: ImageView, rate: Int) {
         when (rate) {
             in 0..33 -> Glide.with(imageView).load(R.drawable.ic_illust_stage_1).into(imageView)
             in 34..66 -> Glide.with(imageView).load(R.drawable.ic_illust_stage_2).into(imageView)
             in 67..99 -> Glide.with(imageView).load(R.drawable.ic_illust_stage_3).into(imageView)
             100 -> Glide.with(imageView).load(R.drawable.ic_illust_stage_4).into(imageView)
         }
+    }
+
+    @BindingAdapter("android:alarmText")
+    @JvmStatic
+    fun TextView.setAlarmText(string: String) {
+        this.text = if (string == "알림 설정") "알림 설정" else setDateFormat(string)
+    }
+
+    private fun setDateFormat(originTime: String): String {
+        Log.d("originTime", originTime) // 2022.01.25 00:04:54
+
+        // 날짜 (2022.01.25)
+        val date =
+            "${originTime[2]}${originTime[3]}.${originTime[5]}${originTime[6]}.${originTime[8]}${originTime[9]}"
+        // 시 (오후 11시 :: 12시간제 적용)
+        val hour = "${originTime[11]}${originTime[12]}".toInt()
+        val newHour = when (hour) {
+            in 0..12 -> " 오전 ${hour}시 "
+            else -> " 오후 ${hour - 12}시 "
+        }
+        // 분 (3분 :: 자릿수 재졍렬을 위한 이중 형변환 사용)
+        val min = "${originTime[14]}${originTime[15]}".toInt().toString() + "분"
+
+        return "$date$newHour$min 알림 예정"
     }
 }
